requirements.txt
webcolors
pandas
pillow